Create a spatially thickened background sample

Same idea as terra::spatSample(weights = weight_raster), but can use a density raster with a different spatial resolution. Can be much quicker than terra::spatSample() for large areas with small resolutions.

Usage

density_sample(
  x,
  dens_rast,
  samp_df,
  boundary,
  mult = 2,
  max_sample = 200,
  verbose = FALSE
)

Arguments

x

spatRaster. Sample will be cell XY values from this raster.

dens_rast

spatRaster. Density raster. This is equivalent to the weights argument of terra::spatSample() but does not have to be at the same resolution as x. Usually made within prep_sdm()

samp_df

Dataframe. Usually built by prep_sdm()

boundary

sf. Boundary within which to sample cells from x

mult

Numeric. In each iteration try sampling mult * sum(samp_df$target) points

max_sample

Numeric. How many iterations to try to meet the density targets in samp_df.

verbose

Logical. Prints how many iterations have been attempted.

Value

Dataframe