Evaluate an SDM

Returns various evaluation metrics from `predicts::pa_evaluate()` and `flexsdm::sdm_eval()`.

evaluate_sdm(m, p_test, b_test, do_gc = FALSE, ...)

Arguments

m

SDM result within `tune_sdm()`

p_test

Presence test data generated within `tune_sdm()`

b_test

Background test data generated within `tune_sdm()`

do_gc

Logical. Run `base::rm(list = ls)` and `base::gc()` at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel. Note, actually usees `rm(list = ls(pattern = "^[^e$]"))`.

...

Passed to both `terra::predict()` and `predicts::pa_evaluate()`

Value

paModelEvaluation (see `predicts::pa_evaluate()`) with extra metrics from `flexsdm::sdm_eval()`: AUC as auc_po_flexsdm; BOYCE as CBI; CBI_rescale (CBI is -1 to 1, CBI_rescale is 0 to 1); and IMAE.

Examples

  library("envSDM")

  out_dir <- file.path(system.file(package = "envSDM"), "examples", "acaule")

  prep <- rio::import(fs::path(out_dir, "prep.rds"))
#> Warning: The `trust` argument of `import()` should be explicit for serialization formats
#> as of rio 1.0.3.
#> ℹ Missing `trust` will be set to FALSE by default for RDS in 2.0.0.
#> ℹ The deprecated feature was likely used in the rio package.
#>   Please report the issue at <https://github.com/gesistsa/rio/issues>.

  presences <- prep$blocks[prep$blocks$pa == 1 & prep$blocks$block != 1, ]
  background <- prep$blocks[prep$blocks$pa == 0 & prep$blocks$block != 1, ]

  evaluate_ex_dir <- fs::path(out_dir, "evaluate_example")

  tune_sdm(out_dir = out_dir
           , algo = "rf"
           , trees = 500
           , mtry = FALSE
           , nodesize = FALSE
           , keep_model = TRUE
           , save_to = evaluate_ex_dir
           )
#> tuning acaule

  tune <- rio::import(fs::path(evaluate_ex_dir, "tune.rds")
                      , setclass = "tibble"
                      )

  evaluate_sdm(tune$m[[1]]
               , p_test = presences
               , b_test = background
               )
#> Error in UseMethod("predict"): no applicable method for 'predict' applied to an object of class "randomForest"