Predict an SDM

Predict an SDM

Usage

predict_sdm(
  this_taxa = NULL,
  out_dir,
  predictors = NULL,
  is_env_pred = TRUE,
  terra_options = NULL,
  doClamp = TRUE,
  limit_to_mcp = TRUE,
  apply_thresh = TRUE,
  force_new = FALSE,
  do_gc = FALSE,
  check_tifs = TRUE,
  ...
)

Arguments

this_taxa

Character. Name of taxa. Used to name outputs. If `NULL`, this will be `basename(dirname(out_dir))`.

out_dir

Character. Name of directory containing model to predict from and into which results will be saved.

predictors

Character. Vector of paths to predictor `.tif` files.

is_env_pred

Logical. Does the naming of the directory and files in `predictors` follow the pattern required by `envRaster::parse_env_tif()`?

terra_options

Passed to `terra::terraOptions()`. e.g. list(memfrac = 0.6)

doClamp

Passed to `terra::predict()` (which then passes as `...` to `fun`). Possibly orphaned from older envSDM?

limit_to_mcp

Logical. If `predict_boundary` exists within `prep` and `limit_to_mcp == TRUE`, the output raster will be limted to within `predict_boundary`.

apply_thresh

Logical. If `TRUE`, an output raster `thresh.tif` will be created using the threshold `mod$e[[1]]@thresholds$max_spec_sens`

force_new

Logical. If outputs already exist, should they be remade?

do_gc

Logical. Run `base::rm(list = ls)` and `base::gc()` at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel.

check_tifs

Logical. Check that any output `.tif` files error on `terra::rast()` and delete them if they do. Useful after crash during predict.

...

Not used.

Value

`invisible(NULL)`. Output `.tif`, log, and optional .png, written to `out_dir`

Examples

  library("envSDM")

  out_dir <- file.path(system.file(package = "envSDM"), "examples")

  data <- file.path(system.file(package = "predicts"), "ex") |>
    fs::dir_ls(regexp = "\\.csv$") |>
    tibble::enframe(name = NULL, value = "path") |>
    dplyr::mutate(taxa = gsub("\\.csv", "", basename(path))
                  , presence = purrr::map(path, rio::import, setclass = "tibble")
                  , presence = purrr::map(presence
                                          , \(x) x |>
                                            dplyr::filter(!is.na(lat)
                                                          , !is.na(lon)
                                                          )
                                          )
                  , out_dir = fs::path(out_dir, taxa)
                  )

  # Run a single model using all the data from the 'best' tune
  purrr::pwalk(list(data$out_dir)
                 , \(a) run_full_sdm(out_dir = a
                                     , metrics_df = envSDM::sdm_metrics
                                     )
                 )
#> run full model for acaule
#> run full model for bradypus

  env_dat <- system.file("ex/bio.tif", package = "predicts")

  # Predict from the best model
  purrr::pwalk(list(data$out_dir)
               , \(a) predict_sdm(out_dir = fs::path(a, "best")
                                  , predictors = env_dat
                                  , is_env_pred = FALSE
                                  , limit_to_mcp = TRUE
                                  , check_tifs = TRUE
                                  )
               )
#> Error in pmap(.l, .f, ..., .progress = .progress): ℹ In index: 1.
#> Caused by error in `predict_sdm()`:
#> ! argument "in_dir" is missing, with no default

  # make .png for each predict
  purrr::walk(data$out_dir
              , \(x) png_from_preds(x
                                    , recurse = 1
                                    , include_blocks = FALSE
                                    )
              )

  # example plots

  # masked to mcp
  purrr::walk(data$out_dir
              , \(x) fs::path(x, "best", "mask.tif") %>%
                terra::rast() %>%
                terra::trim() %>%
                terra::plot()
              )
#> Warning: /home/runner/work/_temp/Library/envSDM/examples/acaule/best/mask.tif: No such file or directory (GDAL error 4)
#> Error in map(.x, .f, ..., .progress = .progress): ℹ In index: 1.
#> Caused by error in `h()`:
#> ! error in evaluating the argument 'x' in selecting a method for function 'plot': error in evaluating the argument 'x' in selecting a method for function 'trim': [rast] file does not exist: /home/runner/work/_temp/Library/envSDM/examples/acaule/best/mask.tif

  # threshold
  purrr::walk(data$out_dir
              , \(x) fs::path(x, "best", "thresh.tif") %>%
                terra::rast() %>%
                terra::trim() %>%
                terra::plot()
              )
#> Warning: /home/runner/work/_temp/Library/envSDM/examples/acaule/best/thresh.tif: No such file or directory (GDAL error 4)
#> Error in map(.x, .f, ..., .progress = .progress): ℹ In index: 1.
#> Caused by error in `h()`:
#> ! error in evaluating the argument 'x' in selecting a method for function 'plot': error in evaluating the argument 'x' in selecting a method for function 'trim': [rast] file does not exist: /home/runner/work/_temp/Library/envSDM/examples/acaule/best/thresh.tif