Prepare for running an SDM

Prepare for running an SDM

prep_sdm(
  this_taxa = NULL,
  out_dir,
  presence,
  pred_limit = TRUE,
  limit_buffer = 0,
  pred_clip = NULL,
  predictors,
  cat_preds = NULL,
  num_bg = 10000,
  prop_abs = "abs",
  use_ecdf = FALSE,
  many_p_prop = 2,
  folds = 5,
  spatial_folds = TRUE,
  min_fold_n = 8,
  min_dens = 0.1,
  use_cuts = 100,
  dens_res = 1000,
  max_samp_iter = 1000,
  low_boost = FALSE,
  save_pngs = TRUE,
  remove_corr = TRUE,
  corr_thresh = 0.9,
  do_gc = FALSE,
  force_new = FALSE
)

Arguments

this_taxa

Character. Name of taxa. Used to name outputs. If `NULL`, this will be `basename(dirname(out_dir))`.

out_dir

Character. Name of directory into which results will be saved.

presence

Dataframe of presences. Needs columns called `lat` (latitude) and `long` (longitude) both in decimal degrees.

pred_limit

Logical. area of interest limits (i.e. MCP around points). Can be `TRUE` (use points to generate a minimum convex polygon to use as a limit), `FALSE` (do not limit) or path to existing sf to use.

limit_buffer

Numeric. Apply this buffer to `pred_limit` (in metres).

pred_clip

sf. Optional sf to clip the pred_limit back to (e.g. to prevent prediction into ocean)

predictors

Character. Vector of paths to predictor `.tif` files. `envRaster::name_env_tif()` is used internally to parse so should follow that naming convention. :( (might work otherwise, but seems unlikely).

cat_preds

Character. Vector of predictor names that are character.

num_bg

Numeric. How many background points?

prop_abs

Character. Is `num_bg` a proportion (`prop`) of the number of records in `presence` or an absolute (`abs`) number?

use_ecdf

Logical. If `TRUE` the density raster is converted via `ras_ecdf <- ecdf(values(target_density)), target_density <- terra::app(target_density, fun = \(i) ras_ecdf(i))`. Creates wider density bands.

many_p_prop

Numeric. Ensure the number of background points is at least `many_p_prop * number of presences`. e.g. If there are more than 5000 presences and num_bg is set at `10000` and `many_p_prop` is `2`, then num_bg will be increased to `many_p_prop * nrow(presences)`

folds

Numeric. How many folds to use in cross validation? Will be adjusted downwards if number of presences do not support `folds * min_fold_n`

spatial_folds

Logical. Use spatial folds? Even if `TRUE`, can resort to non-spatial cv if presences per fold do not meet `min_fold_n` or there are not enough presences to support more than one fold.

min_fold_n

Numeric. Sets both minimum number of presences, and, by default, the minimum number of presences required for a model.

min_dens

Numeric. Usually a little above 0. Sets the minimum density for background points. If zero, large areas of no background points can cause trouble with model predictions into these areas.

use_cuts

Numeric. Used to set number of thresholds in `samp_df`. See `envSDM::density_sample()`.

dens_res

Numeric. Resolution (in metres) of density raster.

max_samp_iter

Numeric. Passed to max_sample argument of `envSDM::density_sample()`.

low_boost

Numeric. Divisor for the lowest density cf the 2nd lowest density. If the lowest density falls below this value, extra background points are added. Helps to stop predicting high probabilities into areas with very few background points.

save_pngs

Logical. Save out a .png of the density raster and spatial blocks

remove_corr

Logical. If TRUE, predictors with high correlation _at presences_ are removed.

corr_thresh

Numeric. Definition of 'high' correlation for `remove_corr`

do_gc

Logical. Run `base::rm(list = ls)` and `base::gc()` at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel.

force_new

Logical. If outputs already exist, should they be remade?

pres_crs

Anything that will return a legitimate crs when passed to the crs attribute of st_transform or st_as_sf