Run an SDM using no cross validation and previously established tune arguments

Run an SDM using no cross validation and previously established tune arguments

Usage

run_full_sdm(
  this_taxa = NULL,
  out_dir,
  use_metric = "combo",
  force_new = FALSE,
  save_to = fs::path(out_dir, use_metric),
  do_gc = FALSE,
  ...
)

Arguments

this_taxa

Character. Name of taxa. Used to name outputs. If `NULL`, this will be `basename(dirname(out_dir))`.

out_dir

Character. Name of directory containing previous tunes and into which results will be saved.

use_metric

Character. Which metric to use to find the 'best' tune arguments from previous tuning results? Default is `combo`, the product of `auc_po`, `CBI_rescale` and `IMAE`. `use_metric` must be `combo` or have been used in the use_metrics argument to `tune_sdm()`.

force_new

Logical. If outputs already exist, should they be remade?

save_to

Character. Name of path to save results. Defaults to `fs::path(out_dir, use_metric)`

do_gc

Logical. Run `base::rm(list = ls)` and `base::gc()` at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel.

...

Passed to `tune_sdm()`

Value

`invisible(NULL)`. `tune.rds` saved into `save_to` directory. log written. `tune.rds` is a data frame but performs poorly, due to list columns , if not imported as a tibble (e.g. via `rio::import(tune.rds, setclass = "tibble"))`)

Examples

inst/examples/predict_sdm_ex.R
#> Error in eval(expr, envir, enclos): object 'inst' not found