Generate .png (map) files for each prediction

Finds any .tif files in pred_dir and writes them to .png files. Includes the retrieval and addition to the map of: various SDM metrics; and the original presence points.

Usage

png_from_preds(
  pred_dir,
  full_run_dir = NULL,
  trim = TRUE,
  force_new = FALSE,
  do_gc = TRUE,
  ...
)

Arguments

trim

Logical. Trim NA values from outside (using terra::trim())

force_new

Logical. If .png file already exists, recreate it?

do_gc

Logical. Run base::rm(list = ls) and base::gc() at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel.

...

Passed to fs::dir_ls()

prep

Character or named list. If character, the path to an existing prep.rds. Otherwise, the result of a call to prep_sdm() with return_val = "object".

full_run

Character or named list. If character, the path to an existing full_run.rds. Otherwise, the result of a call to run_full_sdm() with return_val = "object".

out_dir

Character. Name of directory into which .pngss will be saved. Will be created if it does not exist.

Value

invisible(NULL). Writes .png files with the same file name as any .tif files

Examples