Predict an SDM

Predict an SDM

predict_sdm(
  this_taxa = NULL,
  out_dir,
  predictors = NULL,
  terra_options = NULL,
  doClamp = TRUE,
  limit_to_mcp = TRUE,
  apply_thresh = TRUE,
  force_new = FALSE,
  do_gc = FALSE,
  check_tifs = TRUE,
  ...
)

Arguments

this_taxa

Character. Name of taxa. Used to name outputs. If `NULL`, this will be `basename(dirname(out_dir))`.

out_dir

Character. Name of directory containing model to predict from and into which results will be saved.

predictors

Character. Directory in which predictor `.tif` files live. `envRaster::name_env_tif()` is used internally to parse so should follow that naming convention. :( (might work otherwise, but seems unlikely).

terra_options

Passed to `terra::terraOptions()`. e.g. list(memfrac = 0.6)

doClamp

Passed to `terra::predict()` (which then passes as `...` to `fun`). Possibly orphaned from older envSDM?

limit_to_mcp

Logical. If `predict_boundary` exists within `prep` and `limit_to_mcp == TRUE`, the output raster will be limted to within `predict_boundary`.

apply_thresh

Logical. If `TRUE`, an output raster `thresh.tif` will be created using the threshold `mod$e[[1]]@thresholds$max_spec_sens`

force_new

Logical. If outputs already exist, should they be remade?

do_gc

Logical. Run `base::rm(list = ls)` and `base::gc()` at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel.

check_tifs

Logical. Check that any output `.tif` files error on `terra::rast()` and delete them if they do. Useful after crash during predict.

...

Not used.

Value

`invisible(NULL)`. Output `.tif` files are created.