Predict from SDM

When running predict_sdm, the first predict, saved as 'full.tif' is always to the full extent of the environmental variables even when they extend beyond the predict boundary (this ensures all result can be stacked). The second predict, saved as 'mask.tif' is masked to the same boundary as provided: to the pred_limit argument of prep_sdm; or generated in prep_sdm from the pred_limit, limit_buffer and pred_clip arguments. A threshold raster can also be saved (saved as 'thresh.tif') - see apply_thresh argument.

Usage

predict_sdm(
  prep,
  full_run,
  out_dir,
  predictors = NULL,
  is_env_pred = TRUE,
  terra_options = NULL,
  doClamp = TRUE,
  limit_to_boundary = TRUE,
  apply_thresh = TRUE,
  force_new = FALSE,
  do_gc = FALSE,
  check_tifs = TRUE,
  ...
)

Arguments

prep

Character or named list. If character, the path to an existing prep.rds. Otherwise, the result of a call to prep_sdm() with return_val = "object".

full_run

Character or named list. If character, the path to an existing full_run.rds. Otherwise, the result of a call to run_full_sdm() with return_val = "object".

out_dir

Character. Name of directory into which .tifs will be saved. Will be created if it does not exist.

predictors

Character. Vector of paths to predictor .tif files.

is_env_pred

Logical. Does the naming of the directory and files in predictors follow the pattern required by envRaster::parse_env_tif()?

terra_options

Passed to terra::terraOptions(). e.g. list(memfrac = 0.6)

doClamp

Passed to terra::predict() (which then passes as ... to fun). Possibly orphaned from older envSDM?

limit_to_boundary

Logical. If predict_boundary exists within prep and limit_to_boundary == TRUE, an output raster (mask.tif) will be created within predict_boundary using terra::mask(). Irrespective of limit_to_boundary, full.tif is always created at the full extent of the predictors. Thus all mask.tif files can be 'stacked' as they have the same extent. If needed, limiting the predictions for a taxa to its predict boundary can then be done via terra::trim(mask.tif).

apply_thresh

Logical. If TRUE, an output raster thresh.tif will be created using the maximum of specificity + sensitivity. The threshold value can be accessed within tune.rds as, say, mod <- rio::import("tune.rds") and then mod$e[[1]]@thresholds$max_spec_sens

force_new

Logical. If output files already exist, should they be remade?

do_gc

Logical. Run base::rm(list = ls) and base::gc() at end of function? Useful to keep RAM use down when running SDMs for many, many taxa, especially if done in parallel.

check_tifs

Logical. Check if any output .tif files error on terra::rast() and delete them if they do. Useful after a crash during predict.

...

Passed to terra::predict(). e.g. use for wopt = list(). Also used when masking the full raster back to the mcp (e.g. also passed to terra::mask())

Value

Named list of created .tif files. Names follow the 'full', 'mask' and 'thresh' naming. Output .tif(s), .log, and optional .png, written to out_dir.

Examples