Run an SDM using no cross validation and previously established tune arguments

Run an SDM using no cross validation and previously established tune arguments

run_full_sdm(
  this_taxa = NULL,
  out_dir,
  tune_args = NULL,
  force_new = FALSE,
  save_to = fs::path(out_dir, "best"),
  do_gc = FALSE,
  ...
)

Arguments

this_taxa

Character. Name of taxa. Used to name outputs. If `NULL`, this will be `basename(dirname(out_dir))`.

out_dir

Character. Name of directory containing previous tunes and into which results will be saved.

tune_args

List with element `algo` and elements: `fc`, `rm`, `trees`, `mtry` and `nodesize` as appropriate to the `algo.` If supplied as `NULL` an attempt will be made to load 'best' tune from `evaluation.csv` in `out_dir`

force_new

Logical. If outputs already exist, should they be remade?

save_to

Character. Name of path to save results. Defaults to `fs::path(out_dir, "best")`

do_gc

Logical. Run `base::rm(list = ls)` and `base::gc()` at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel.

...

Passed to `tune_sdm()`

Value

`invisible(NULL)`. `tune.rds` saved into `save_to` directory. log written. `tune.rds` is a data frame but performs poorly, due to list columns , if not imported as a tibble (e.g. via `rio::import(tune.rds, setclass = "tibble"))`)

Examples

inst/examples/predict_sdm_ex.R
#> Error in eval(expr, envir, enclos): object 'inst' not found