Tune, and evaluate, species distribution models

Tune, and evaluate, species distribution models

tune_sdm(
  this_taxa = NULL,
  out_dir,
  algo = c("all", "maxnet", "bioclim", "envelope", "rf"),
  fc = "auto_feature",
  limit_p = FALSE,
  rm = seq(1, 6, 0.5),
  trees = c(500, 1000, 2000),
  mtry = TRUE,
  nodesize = c(1, 2),
  keep_model = FALSE,
  best_run = FALSE,
  metrics_df = envSDM::sdm_metrics,
  do_gc = TRUE,
  force_new = FALSE,
  save_to = out_dir,
  ...
)

Arguments

this_taxa

Character. Name of taxa. Used to name outputs. If `NULL`, this will be `basename(dirname(out_dir))`.

out_dir

Character. Name of directory into which results will be saved.

algo

Character. Name of algorithm to use.

fc

Character. Used to generate levels of `classes` argument to `maxnet::maxnet()` that are tuned.

limit_p

`TRUE`, `FALSE` or number of predictor variables above which to limit the use of `p` in the classes argument used in `maxnet::maxnet()`. Useful with many predictor variables when it becomes unwieldy to generate interactions for all predictors.

rm

Numeric. Used to generate levels of `regmult` argument to `maxnet::maxnet()` that are tuned.

trees

Used to generate the levels of `ntree` argument to `randomForest::randomForest()` that are tuned. `TRUE` (tune with default `trees`), `FALSE` (don't tune `trees`) or numeric (the `trees` values to tune with).

mtry

Used to generate the levels of `mtry` argument to `randomForest::randomForest()` that are tuned. `TRUE` (tune with default `mtry`), `FALSE` (only use default `randomForest::randomForest()` `mtry`) or numeric (the `mtry` values to tune with).

nodesize

Used to generate the levels of `nodesize` argument to `randomForest::randomForest()` that are tuned. `TRUE` (tune with default `nodesize`), `FALSE` (only use default `randomForest::randomForest()` `nodesize`) or numeric (the `nodesize` values to tune with).

keep_model

Logical. If `TRUE` the model results will be appended as a list column in the returned tibble (as column `m`)

best_run

Logical. If `TRUE` this alters the behaviour of the `tune_sdm()` by, well, not tuning. :). Sets all blocks to the same value so no cross-validation.

metrics_df

Dataframe. Defines which metrics to use when deciding on 'good' SDMs.

do_gc

Logical. Run `base::rm(list = ls)` and `base::gc()` at end of function? Useful when running SDMs for many, many taxa, especially if done in parallel.

force_new

Logical. If outputs already exist, should they be remade?

save_to

Character. Name of path to save results. Defaults to out_dir which works if not `best_run`. For a best run from `run_full_sdm()`, this will default to `fs::path(out_dir, "best")`. Otherwise set as anything.

...

Passed to `evaluate_sdm()`